Replace lexima.vim with nvim-autopairs
There's a weird interaction between lexima and `formatoptions+=r` that
I haven't been able to figure out the root cause quickly. Essentially:

    if condition {
        // commentstring|
    }
                        ^- Normal typing until the cursor reaches here

Pressing <CR> leads to the `r` behaviour of prefixing the line with
`&commentstring`, but leaving insert mode leads to this:

    if condition {
        // commentstring
        //}

Since I'm not keen on re-learning vimscript for the hudredth time just
for fixing this, I just switched to autopairs...

Should also make my life easier if I ever follow through with the
thought of adding an auto-complete plugin.