Remove wezterm config
Been back on Konsole for a while now. Switched because there's a bug
where I shrinks to a tiny square when I switch displays with different
scaling settings.

Dunno if it still happens, dunno if it were my fault smh. I haven't
missed it at all, so away it goes.